+++ changelog.md
@@ -4,6 +4,7 @@
 
 - new: server: count the bytes we rec/send
 - new: CryptUtils
+- serial: SSL -> CryptUtils
 
 ## Version 4.7.2